Added Permission control on QA - Assessments

+++ resources/views/assessments/index.blade.php
@@ -38,20 +38,24 @@
             @if(count($assessments) > 0)
             <div id="toolbar" class="shadow-eff1">
                 <div class="btn-group">
-                    <button id="sidebarCollapse" class="btn btn-default" data-toggle="offcanvas">
-                        <i class="glyphicon glyphicon-align-left"></i>
-                        <span>Filters</span>
-                    </button>
-                    <button id="item-create" type="button" class="btn btn-sham" data-wenk="Add new" data-wenk-pos="bottom">
-                        <i class="glyphicon glyphicon-plus"></i> Add New
-                    </button>
+                    @if($allowedActions->contains('List'))
+                        <button id="sidebarCollapse" class="btn btn-default" data-toggle="offcanvas">
+                            <i class="glyphicon glyphicon-align-left"></i>
+                            <span>Filters</span>
+                        </button>
+                    @endif
+                    @if($allowedActions->contains('Create'))
+                        <button id="item-create" type="button" class="btn btn-sham" data-wenk="Add new" data-wenk-pos="bottom">
+                            <i class="glyphicon glyphicon-plus"></i> Add New
+                        </button>
+                    @endif
                 </div>
             </div>
             @endif
             <div class="table-responsive">
             @if(count($assessments) == 0)
                 <h4 class="text-center">Its a bit empty here. You may click <a href="javascript:;" class="text-primary item-create">here</a> to add a new assessment</h4>
-            @else
+            @elseif($allowedActions->contains('List'))
                 <table id="new-table" data-toggle="table">
                     <thead>
                         <tr>
@@ -71,12 +75,16 @@
 
                             <td data-html2canvas-ignore="true">
                                 <div class="btn-group btn-group-xs" role="group">
-                                    <a href="#modal-text" data-wenk="Edit" class="b-n b-n-r bg-transparent item-edit" onclick="editFullPage('{{$assessment->id}}', event)">
-                                        <i class="glyphicon glyphicon-edit text-primary"></i>
-                                    </button>
-                                    <a href="#!" class="b-n b-n-r bg-transparent item-remove" data-wenk="Remove" onclick="deleteForm('{{$assessment->id}}')">
-                                        <i class="glyphicon glyphicon-remove text-danger"></i>
-                                    </button>
+                                    @if($allowedActions->contains('Write'))
+                                        <a href="#modal-text" data-wenk="Edit" class="b-n b-n-r bg-transparent item-edit" onclick="editFullPage('{{$assessment->id}}', event)">
+                                            <i class="glyphicon glyphicon-edit text-primary"></i>
+                                            </a>
+                                    @endif
+                                    @if($allowedActions->contains('Delete'))
+                                            <a href="#!" class="b-n b-n-r bg-transparent item-remove" data-wenk="Remove" onclick="deleteForm('{{$assessment->id}}')">
+                                                <i class="glyphicon glyphicon-remove text-danger"></i>
+                                            </a>
+                                    @endif
                                 </div>
                             </td>
                         </tr>
